start moving neo-express over to using neo2 dll
