+++ .vscode/launch.json
@@ -1,19 +1,17 @@
 {
-   // Use IntelliSense to find out which attributes exist for C# debugging
-   // Use hover for the description of the existing attributes
-   // For further information visit https://github.com/OmniSharp/omnisharp-vscode/blob/master/debugger-launchjson.md
-   "version": "0.2.0",
-   "configurations": [
+    // Use IntelliSense to learn about possible attributes.
+    // Hover to view descriptions of existing attributes.
+    // For more information, visit: https://go.microsoft.com/fwlink/?linkid=830387
+    "version": "0.2.0",
+    "configurations": [
         {
             "name": ".NET Core Launch (console)",
             "type": "coreclr",
             "request": "launch",
-            "preLaunchTask": "build",
-            // If you have changed target frameworks, make sure to update the program path.
-            "program": "${workspaceFolder}/src/neo-express/bin/Debug/netcoreapp3.1/neo-express.dll",
-            "args": [],
-            "cwd": "${workspaceFolder}/src/neo-express",
-            // For more information about the 'console' field, see https://aka.ms/VSCode-CS-LaunchJson-Console
+            "preLaunchTask": "build neo2",
+            "program": "${workspaceFolder}/src/neo2/bin/Debug/netcoreapp3.1/neo2.dll",
+            "args": ["run"],
+            "cwd": "${workspaceFolder}/src/neo2",
             "console": "internalConsole",
             "stopAtEntry": false
         },
