--- src/neo-express/BlockchainOperations.cs
@@ -1,338 +1,0 @@
-﻿using Neo;
-using NeoExpress.Models;
-using NeoExpress.Node;
-using NeoExpress.Persistence;
-using Newtonsoft.Json;
-using System;
-using System.Collections.Generic;
-using System.Diagnostics;
-using System.IO;
-using System.Linq;
-using System.Net;
-using System.Text;
-using System.Threading;
-using System.Threading.Tasks;
-
-namespace NeoExpress
-{
-    internal static class BlockchainOperations
-    {
-        public static ExpressChain CreateBlockchain(int count)
-        {
-            var wallets = new List<(DevWallet wallet, Neo.Wallets.WalletAccount account)>(count);
-
-            ushort GetPortNumber(int index, ushort portNumber) => (ushort)((49000 + (index * 1000)) + portNumber);
-
-            try
-            {
-                for (var i = 1; i <= count; i++)
-                {
-                    var wallet = new DevWallet($"node{i}");
-                    var account = wallet.CreateAccount();
-                    account.IsDefault = true;
-                    wallets.Add((wallet, account));
-                }
-
-                var keys = wallets.Select(t => t.account.GetKey().PublicKey).ToArray();
-
-                var contract = Neo.SmartContract.Contract.CreateMultiSigContract((keys.Length * 2 / 3) + 1, keys);
-
-                foreach (var (wallet, account) in wallets)
-                {
-                    var multiSigContractAccount = wallet.CreateAccount(contract, account.GetKey());
-                    multiSigContractAccount.Label = "MultiSigContract";
-                }
-
-                // 49152 is the first port in the "Dynamic and/or Private" range as specified by IANA
-                // http://www.iana.org/assignments/port-numbers
-                var nodes = new List<ExpressConsensusNode>(count);
-                for (var i = 0; i < count; i++)
-                {
-                    nodes.Add(new ExpressConsensusNode()
-                    {
-                        TcpPort = GetPortNumber(i, 333),
-                        WebSocketPort = GetPortNumber(i, 334),
-                        RpcPort = GetPortNumber(i, 332),
-                        Wallet = wallets[i].wallet.ToExpressWallet()
-                    });
-                }
-
-                return new ExpressChain()
-                {
-                    Magic = ExpressChain.GenerateMagicValue(),
-                    ConsensusNodes = nodes,
-                };
-            }
-            finally
-            {
-                foreach (var (wallet, _) in wallets)
-                {
-                    wallet.Dispose();
-                }
-            }
-        }
-
-        public static void ExportBlockchain(ExpressChain chain, string folder, string password, Action<string> writeConsole)
-        {
-            void WriteNodeConfigJson(ExpressConsensusNode _node, string walletPath)
-            {
-                using (var stream = File.Open(Path.Combine(folder, $"{_node.Wallet.Name}.config.json"), FileMode.Create, FileAccess.Write))
-                using (var writer = new JsonTextWriter(new StreamWriter(stream)) { Formatting = Formatting.Indented })
-                {
-                    writer.WriteStartObject();
-                    writer.WritePropertyName("ApplicationConfiguration");
-                    writer.WriteStartObject();
-
-                    writer.WritePropertyName("Paths");
-                    writer.WriteStartObject();
-                    writer.WritePropertyName("Chain");
-                    writer.WriteValue("Chain_{0}");
-                    writer.WritePropertyName("Index");
-                    writer.WriteValue("Index_{0}");
-                    writer.WriteEndObject();
-
-                    writer.WritePropertyName("P2P");
-                    writer.WriteStartObject();
-                    writer.WritePropertyName("Port");
-                    writer.WriteValue(_node.TcpPort);
-                    writer.WritePropertyName("WsPort");
-                    writer.WriteValue(_node.WebSocketPort);
-                    writer.WriteEndObject();
-
-                    writer.WritePropertyName("RPC");
-                    writer.WriteStartObject();
-                    writer.WritePropertyName("BindAddress");
-                    writer.WriteValue("127.0.0.1");
-                    writer.WritePropertyName("Port");
-                    writer.WriteValue(_node.RpcPort);
-                    writer.WritePropertyName("SslCert");
-                    writer.WriteValue("");
-                    writer.WritePropertyName("SslCertPassword");
-                    writer.WriteValue("");
-                    writer.WriteEndObject();
-
-                    writer.WritePropertyName("UnlockWallet");
-                    writer.WriteStartObject();
-                    writer.WritePropertyName("Path");
-                    writer.WriteValue(walletPath);
-                    writer.WritePropertyName("Password");
-                    writer.WriteValue(password);
-                    writer.WritePropertyName("StartConsensus");
-                    writer.WriteValue(true);
-                    writer.WritePropertyName("IsActive");
-                    writer.WriteValue(true);
-                    writer.WriteEndObject();
-
-                    writer.WriteEndObject();
-                    writer.WriteEndObject();
-                }
-            }
-
-            void WriteProtocolJson()
-            {
-                using (var stream = File.Open(Path.Combine(folder, "protocol.json"), FileMode.Create, FileAccess.Write))
-                using (var writer = new JsonTextWriter(new StreamWriter(stream)) { Formatting = Formatting.Indented })
-                {
-                    writer.WriteStartObject();
-                    writer.WritePropertyName("ProtocolConfiguration");
-                    writer.WriteStartObject();
-
-                    writer.WritePropertyName("Magic");
-                    writer.WriteValue(chain.Magic);
-                    writer.WritePropertyName("AddressVersion");
-                    writer.WriteValue(23);
-                    writer.WritePropertyName("SecondsPerBlock");
-                    writer.WriteValue(15);
-
-                    writer.WritePropertyName("StandbyValidators");
-                    writer.WriteStartArray();
-                    for (int i = 0; i < chain.ConsensusNodes.Count; i++)
-                    {
-                        var account = DevWalletAccount.FromExpressWalletAccount(chain.ConsensusNodes[i].Wallet.DefaultAccount);
-                        var key = account.GetKey();
-                        if (key != null)
-                        {
-                            writer.WriteValue(key.PublicKey.EncodePoint(true).ToHexString());
-                        }
-                    }
-                    writer.WriteEndArray();
-
-                    writer.WritePropertyName("SeedList");
-                    writer.WriteStartArray();
-                    foreach (var node in chain.ConsensusNodes)
-                    {
-                        writer.WriteValue($"{IPAddress.Loopback}:{node.TcpPort}");
-                    }
-                    writer.WriteEndArray();
-
-                    writer.WriteEndObject();
-                    writer.WriteEndObject();
-                }
-            }
-
-            for (var i = 0; i < chain.ConsensusNodes.Count; i++)
-            {
-                var node = chain.ConsensusNodes[i];
-                writeConsole($"Exporting {node.Wallet.Name} Conensus Node wallet");
-
-                var walletPath = Path.Combine(folder, $"{node.Wallet.Name}.wallet.json");
-                if (File.Exists(walletPath))
-                {
-                    File.Delete(walletPath);
-                }
-
-                ExportWallet(node.Wallet, walletPath, password);
-                WriteNodeConfigJson(node, walletPath);
-            }
-
-            WriteProtocolJson();
-        }
-
-        public static ExpressWallet CreateWallet(string name)
-        {
-            using (var wallet = new DevWallet(name))
-            {
-                var account = wallet.CreateAccount();
-                account.IsDefault = true;
-                return wallet.ToExpressWallet();
-            }
-        }
-
-        public static void ExportWallet(ExpressWallet wallet, string filename, string password)
-        {
-            var devWallet = DevWallet.FromExpressWallet(wallet);
-            devWallet.Export(filename, password);
-        }
-
-        public static (byte[] signature, byte[] publicKey) Sign(ExpressWalletAccount account, byte[] data)
-        {
-            var devAccount = DevWalletAccount.FromExpressWalletAccount(account);
-
-            var key = devAccount.GetKey();
-            if (key == null)
-                throw new InvalidOperationException();
-
-            var publicKey = key.PublicKey.EncodePoint(false).AsSpan().Slice(1).ToArray();
-            var signature = Neo.Cryptography.Crypto.Default.Sign(data, key.PrivateKey, publicKey);
-            return (signature, key.PublicKey.EncodePoint(true));
-        }
-
-        private const string ADDRESS_FILENAME = "ADDRESS.neo-express";
-
-        private static string GetAddressFilePath(string directory) =>
-            Path.Combine(directory, ADDRESS_FILENAME);
-
-        public static void CreateCheckpoint(ExpressChain chain, string blockChainStoreDirectory, string checkPointFileName)
-        {
-            using var db = new RocksDbStore(blockChainStoreDirectory);
-            CreateCheckpoint(db, checkPointFileName, chain.Magic, chain.ConsensusNodes[0].Wallet.DefaultAccount.ScriptHash);
-        }
-
-        public static void CreateCheckpoint(RocksDbStore db, string checkPointFileName, long magic, string scriptHash)
-        {
-            string tempPath;
-            do
-            {
-                tempPath = Path.Combine(Path.GetTempPath(), Path.GetRandomFileName());
-            }
-            while (Directory.Exists(tempPath));
-
-            try
-            {
-                db.CheckPoint(tempPath);
-
-                using (var stream = File.OpenWrite(GetAddressFilePath(tempPath)))
-                using (var writer = new StreamWriter(stream))
-                {
-                    writer.WriteLine(magic);
-                    writer.WriteLine(scriptHash);
-                }
-
-                if (File.Exists(checkPointFileName))
-                {
-                    throw new InvalidOperationException(checkPointFileName + " checkpoint file already exists");
-                }
-                System.IO.Compression.ZipFile.CreateFromDirectory(tempPath, checkPointFileName);
-            }
-            finally
-            {
-                Directory.Delete(tempPath, true);
-            }
-        }
-
-        public static void RestoreCheckpoint(ExpressChain chain, string chainDirectory, string checkPointDirectory)
-        {
-            var node = chain.ConsensusNodes[0];
-            ValidateCheckpoint(checkPointDirectory, chain.Magic, node.Wallet.DefaultAccount);
-
-            var addressFile = GetAddressFilePath(checkPointDirectory);
-            if (!File.Exists(addressFile))
-            {
-                File.Delete(addressFile);
-            }
-
-            Directory.Move(checkPointDirectory, chainDirectory);
-        }
-
-        private static void ValidateCheckpoint(string checkPointDirectory, long magic, ExpressWalletAccount account)
-        {
-            var addressFile = GetAddressFilePath(checkPointDirectory);
-            if (!File.Exists(addressFile))
-            {
-                throw new Exception("Invalid Checkpoint");
-            }
-
-            long checkPointMagic;
-            string scriptHash;
-            using (var stream = File.OpenRead(addressFile))
-            using (var reader = new StreamReader(stream))
-            {
-                checkPointMagic = long.Parse(reader.ReadLine() ?? string.Empty);
-                scriptHash = reader.ReadLine() ?? string.Empty;
-            }
-
-            if (magic != checkPointMagic || scriptHash != account.ScriptHash)
-            {
-                throw new Exception("Invalid Checkpoint");
-            }
-        }
-
-        public static void PreloadGas(string directory, ExpressChain chain, int index, uint preloadGasAmount, TextWriter writer, CancellationToken cancellationToken)
-        {
-            if (!chain.InitializeProtocolSettings())
-            {
-                throw new Exception("could not initialize protocol settings");
-            }
-            var node = chain.ConsensusNodes[index];
-            using var store = new RocksDbStore(directory);
-            NodeUtility.Preload(preloadGasAmount, store , node, writer, cancellationToken);
-        }
-
-        public static Task RunBlockchainAsync(string directory, ExpressChain chain, int index, uint secondsPerBlock, TextWriter writer, CancellationToken cancellationToken)
-        {
-            if (!chain.InitializeProtocolSettings(secondsPerBlock))
-            {
-                throw new Exception("could not initialize protocol settings");
-            }
-            var node = chain.ConsensusNodes[index];
-
-#pragma warning disable IDE0067 // NodeUtility.RunAsync disposes the store when it's done
-            return NodeUtility.RunAsync(new RocksDbStore(directory), node, writer, cancellationToken);
-#pragma warning restore IDE0067 // Dispose objects before losing scope
-        }
-
-        public static Task RunCheckpointAsync(string directory, ExpressChain chain, uint secondsPerBlock, TextWriter writer, CancellationToken cancellationToken)
-        {
-            if (!chain.InitializeProtocolSettings(secondsPerBlock))
-            {
-                throw new Exception("could not initialize protocol settings");
-            }
-            var node = chain.ConsensusNodes[0];
-            ValidateCheckpoint(directory, chain.Magic, node.Wallet.DefaultAccount);
-
-#pragma warning disable IDE0067 // NodeUtility.RunAsync disposes the store when it's done
-            return NodeUtility.RunAsync(new CheckpointStore(directory), node, writer, cancellationToken);
-#pragma warning restore IDE0067 // Dispose objects before losing scope
-        }
-    }
-}
