+++ src/neo-express/Commands/ExportCommand.cs
@@ -19,7 +19,8 @@ namespace NeoExpress.Commands
                 var (chain, _) = Program.LoadExpressChain(Input);
                 var password = Prompt.GetPassword("Input password to use for exported wallets");
 
-                BlockchainOperations.ExportBlockchain(chain, Directory.GetCurrentDirectory(), password, msg => console.WriteLine(msg));
+                var blockchainOperations = new NeoExpress.Neo2.BlockchainOperations();
+                blockchainOperations.ExportBlockchain(chain, Directory.GetCurrentDirectory(), password, msg => console.WriteLine(msg));
 
                 return 0;
             }
