+++ src/neo-express/Commands/TransferCommand.cs
@@ -6,7 +6,6 @@ using System.Linq;
 using System.Threading.Tasks;
 using System.Collections.Generic;
 using NeoExpress.Models;
-using NeoExpress.Node;
 
 namespace NeoExpress.Commands
 {
