+++ src/neo-express/Commands/WalletCommand.Export.cs
@@ -52,7 +52,8 @@ namespace NeoExpress.Commands
                     else
                     {
                         var password = Prompt.GetPassword("Input password to use for exported wallet");
-                        BlockchainOperations.ExportWallet(wallet, output, password);
+                        var blockchainOperations = new NeoExpress.Neo2.BlockchainOperations();
+                        blockchainOperations.ExportWallet(wallet, output, password);
                         console.WriteLine($"{Name} privatenet wallet exported to {output}");
                     }
