+++ src/neo-express/default.neo-express.json
@@ -1,0 +1,41 @@
+{
+  "magic": 1661073876,
+  "consensus-nodes": [
+    {
+      "tcp-port": 49333,
+      "ws-port": 49334,
+      "rpc-port": 49332,
+      "wallet": {
+        "name": "node1",
+        "accounts": [
+          {
+            "private-key": "9f885b4f0f4ea533081412f18447b7813d5aab1aec476e8276872965b9dacfac",
+            "script-hash": "AHirNQuDyJu6SbCY9ZTiK7USHDMxjZqMrq",
+            "label": null,
+            "is-default": true,
+            "contract": {
+              "script": "2103418b26c773f9aaa91140f0ebe5f75e64eec64ee3d0528444312fd782ef1a1eafac",
+              "parameters": [
+                "Signature"
+              ]
+            }
+          },
+          {
+            "private-key": "9f885b4f0f4ea533081412f18447b7813d5aab1aec476e8276872965b9dacfac",
+            "script-hash": "AVVWYBVH7pZuGGNcn3rXAQjmzaP8artEzQ",
+            "label": "MultiSigContract",
+            "is-default": false,
+            "contract": {
+              "script": "512103418b26c773f9aaa91140f0ebe5f75e64eec64ee3d0528444312fd782ef1a1eaf51ae",
+              "parameters": [
+                "Signature"
+              ]
+            }
+          }
+        ]
+      }
+    }
+  ],
+  "wallets": [],
+  "contracts": []
+}
