--- src/neo2/Commands/WalletCommand.cs
@@ -1,16 +1,0 @@
-﻿using McMaster.Extensions.CommandLineUtils;
-
-namespace NeoExpress.Neo2.Commands
-{
-    [Command("wallet")]
-    [Subcommand(typeof(Create), typeof(Delete), typeof(Export), typeof(List))]
-    internal partial class WalletCommand
-    {
-        private int OnExecute(CommandLineApplication app, IConsole console)
-        {
-            console.WriteError("You must specify at a subcommand.");
-            app.ShowHelp();
-            return 1;
-        }
-    }
-}
